[ft_irc] refactor: organizing code, functions and files
